create views_hw33-34.py move all old views from views.py to views_hw33-34.py
